+++ hutool-poi/src/test/java/cn/hutool/poi/excel/test/ExcelWriteTest.java
@@ -102,6 +102,40 @@ public class ExcelWriteTest {
 		// 关闭writer，释放内存
 		writer.close();
 	}
+	
+	@Test
+	@Ignore
+	public void mergeTest2() {
+		Map<String, Object> row1 = new LinkedHashMap<>();
+		row1.put("姓名", "张三");
+		row1.put("年龄", 23);
+		row1.put("成绩", 88.32);
+		row1.put("是否合格", true);
+		row1.put("考试日期", DateUtil.date());
+		
+		Map<String, Object> row2 = new LinkedHashMap<>();
+		row2.put("姓名", "李四");
+		row2.put("年龄", 33);
+		row2.put("成绩", 59.50);
+		row2.put("是否合格", false);
+		row2.put("考试日期", DateUtil.date());
+		
+		ArrayList<Map<String, Object>> rows = CollUtil.newArrayList(row1, row2);
+
+		// 通过工具类创建writer
+		ExcelWriter writer = ExcelUtil.getWriter("e:/writeMapTest.xlsx");
+		// 合并单元格后的标题行，使用默认标题样式
+		try {
+			writer.merge(row1.size() - 1, "一班成绩单");
+		} catch (Exception e) {
+			e.printStackTrace();
+		}
+		
+		// 一次性写出内容，使用默认样式，强制输出标题
+		writer.write(rows, true);
+		// 关闭writer，释放内存
+		writer.close();
+	}
 
 	@Test
 	@Ignore